==> notebooks/5.ipynb <==
# Machine Learning Modeling

import pandas as pd
import numpy as np
import matplotlib.pyplot as plot
from sklearn import metrics
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression

df = pd.read_pickle("../data/production/imputed_dataset.pickle")

x = df.iloc[:, :-1]
y = df.iloc[:, -1]

x.head(5)

y.head(5)

xTest, xTrain, yTest, yTrain = train_test_split(x, y, test_size=0.7, random_state=0)

## Logistic Regression

logreg = LogisticRegression()
logreg.fit(xTrain, yTrain)

yPred = logreg.predict(xTest)
print('Accuracy of logistic regression classifier on test set: {:.2f}'.format(logreg.score(xTest, yTest)))

confusion = confusion_matrix(yTest, yPred)
print(confusion)

print(classification_report(yTest, yPred))

import pickle
with open("../models/logreg.pickle", "wb") as dbfile:
    pickle.dump(logreg,dbfile)
