==> notebooks/8.ipynb <==
# Model Visualization

""" Notebook Setup/Imports
"""
!pip install -U altair vega_datasets vega pandas-profiling
# Project path setup
from pathlib import Path
project_dir = Path(__name__).resolve().parents[1]

# Library import
import pandas as pd
import numpy as np

# Print strings as markdown
from IPython.display import Markdown, HTML

# Data variables from previous notebook
models_df = pd.read_pickle("../models/predictions.pickle")

# Plotting libraries
import altair as alt
from vega_datasets import data
alt.renderers.enable('default')

from pandas_profiling import ProfileReport

counties = alt.topo_feature(data.us_10m.url, 'counties')
source = models_df.reset_index()

alt.Chart(counties).mark_geoshape().encode(
    color='USDA Model:Q'
).transform_lookup(
    lookup='id',
    from_=alt.LookupData(source, 'index', ['USDA Model'])
).project(
    type='albersUsa'
).properties(
    width=500,
    height=300
).save("usda.html")

models_df

alt.Chart(counties).mark_geoshape().encode(
    color='Logistic Regression:Q'
).transform_lookup(
    lookup='id',
    from_=alt.LookupData(source, 'index', ['Logistic Regression'])
).project(
    type='albersUsa'
).properties(
    width=500,
    height=300
).save("logreg.html")

alt.Chart(counties).mark_geoshape().encode(
    color='Adamax:Q'
).transform_lookup(
    lookup='id',
    from_=alt.LookupData(source, 'index', ['Adamax'])
).project(
    type='albersUsa'
).properties(
    width=500,
    height=300
).save("adamax.html")

alt.Chart(counties).mark_geoshape().encode(
    color='Nadam:Q'
).transform_lookup(
    lookup='id',
    from_=alt.LookupData(source, 'index', ['Nadam'])
).project(
    type='albersUsa'
).properties(
    width=500,
    height=300
).save("nadam.html")

profile = ProfileReport(models_df)

profile.to_file("../docs/_static/final_profile.html")
